webshop/requirements.txt
asgiref==3.7.2
colorlog==6.7.0
Django==4.2.3
django-debug-toolbar==4.1.0
django-log-request-id==2.1.0
django-taggit==4.0.0
djangorestframework==3.14.0
idna==3.4
numpy==1.25.1
Pillow==10.0.0
python-dotenv==1.0.0
pytz==2023.3
PyWavelets==1.4.1
scipy==1.11.1
six==1.16.0
sqlparse==0.4.4
